% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{DEG_list}
\alias{DEG_list}
\title{DEG_list}
\format{
A list containing data frames with 1645 rows and 4 variables and with 1711 rows and 4 variables
}
\usage{
data(DEG_list)
}
\description{
example data for vennd, ORA, GeneSetConversion, and AutoExtraction
}
