% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_RNAseq_analysis.R
\name{DEG_overview}
\alias{DEG_overview}
\title{Visualization of pairwise DEG analysis}
\usage{
DEG_overview(
  Count_matrix,
  DEG_result,
  Type = "EBseq",
  Species,
  Cond_1 = 3,
  Cond_2 = 3,
  fdr = 0.05,
  fc = 2
)
}
\arguments{
\item{Count_matrix}{Count matrix}

\item{DEG_result}{result of DEG analysis}

\item{Type}{one of "EBseq" or "DEseq2"}

\item{Species}{Species}

\item{Cond_1}{Sumple number of condition_1}

\item{Cond_2}{Sumple number of condition_2}

\item{fdr}{Accepted false discovery rate for considering genes as differentially expressed}

\item{fc}{the fold change threshold. Only genes with a fold change >= fc and padj <= fdr are considered as significantly differentially expressed.}
}
\description{
Visualization of pairwise DEG analysis
}
