% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{deseq2}
\alias{deseq2}
\title{deseq2}
\usage{
deseq2(Row_count_matrix)
}
\arguments{
\item{Row_count_matrix}{Row Count matrix txt file (Not normalized count matrix)}
}
\description{
deseq2
}
\examples{
library(rnaseqviewer)

data(Row_count_data)
write.table(Row_count_data, file = "Row_count_data.txt", sep = "\t", quote = FALSE)
deseq2("Row_count_data.txt")

}
\references{
Love, M.I., Huber, W., Anders, S. Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2 Genome Biology 15(12):550 (2014)
}
