% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR.R
\name{edger}
\alias{edger}
\title{edgeR}
\usage{
edger(Row_count_matrix, method = "BH")
}
\arguments{
\item{Row_count_matrix}{Row Count matrix txt file (Not normalized count matrix)}

\item{method}{BH, Qvalue, or IHW}
}
\description{
edgeR
}
\examples{
library(rnaseqviewer)

data(Row_count_data)
write.table(Row_count_data, file = "Row_count_data.txt", sep = "\t", quote = FALSE)
edger("Row_count_data.txt")

}
\references{
Robinson MD, McCarthy DJ and Smyth GK (2010). edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139-140

Nikolaos Ignatiadis, Bernd Klaus, Judith Zaugg and Wolfgang Huber (2016): Data-driven hypothesis weighting increases detection power in genome-scale multiple testing. Nature Methods 13:577, doi: 10.1038/nmeth.3885

John D. Storey, Andrew J. Bass, Alan Dabney and David Robinson (2021). qvalue: Q-value estimation for false discovery rate control. R package version 2.26.0. http://github.com/jdstorey/qvalue
}
