% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_heatmap.R
\name{int_heatmap}
\alias{int_heatmap}
\title{Integrated heatmap}
\usage{
int_heatmap(Count_matrix_dir, Gene_set, pre_zscoring = T)
}
\arguments{
\item{Count_matrix_dir}{Directory including count matrix txt files}

\item{Gene_set}{gene set txt file}

\item{pre_zscoring}{z-scoring before integration of data set}
}
\description{
Integrated heatmap
}
\examples{
library(rnaseqviewer)

data(DEG_list)
dir.create("DEG_list")
write.table(DEG_list[1], file = "DEG_list/dataset1.txt", sep = "\t", quote = FALSE)
write.table(DEG_list[2], file = "DEG_list/dataset2.txt", sep = "\t", quote = FALSE)
vennd("DEG_list")

dir.create("count_list")
data(Row_count_data)
data(Row_count_data2)
write.table(Row_count_data, file = "count_list/data1.txt", sep = "\t", quote = FALSE)
write.table(Row_count_data2, file = "count_list/data2.txt", sep = "\t", quote = FALSE)

int_heatmap(Count_matrix_dir = "count_list",
            Gene_set = "DEG_list/group_lists/dataset1:dataset2.txt")

}
\references{
R. Gentleman, V. Carey, W. Huber and F. Hahne (2021). genefilter: methods for filtering genes from high-throughput experiments. R package version 1.72.1.

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.
}
