% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_RNAseq_analysis.R
\name{kmeansClustring}
\alias{kmeansClustring}
\title{Kmeans clustering analysis}
\usage{
kmeansClustring(Count_matrix, km, km_repeats = 10000, basemean = 0)
}
\arguments{
\item{Count_matrix}{Directory of count matrix txt file}

\item{km}{number of k-means clustering}

\item{km_repeats}{number of k-means runs to get a consensus k-means clustering}

\item{basemean}{basemean threshold}
}
\description{
Kmeans clustering analysis
}
\references{
R. Gentleman, V. Carey, W. Huber and F. Hahne (2021). genefilter: methods for filtering genes from high-throughput experiments. R package version 1.72.1.

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.
}
