% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comparison_EBseq_viewer.R
\name{multiDEG_overview}
\alias{multiDEG_overview}
\title{Visualization of multiple comparison DEG analysis}
\usage{
multiDEG_overview(
  Normalized_count_matrix,
  EBseq_result,
  EBseq_condmeans,
  Species = NULL,
  fdr = 0.05,
  fc = 2,
  basemean = 0
)
}
\arguments{
\item{Normalized_count_matrix}{Count matrix txt file (e.g. TPM count matrix.txt)}

\item{EBseq_result}{result txt file of EBseq analysis}

\item{EBseq_condmeans}{Condmeands txt file from EBseq analysis}

\item{Species}{Species}

\item{fdr}{Accepted false discovery rate for considering genes as differentially expressed}

\item{fc}{the fold change threshold. Only genes with a fold change >= fc and padj <= fdr are considered as significantly differentially expressed.}

\item{basemean}{basemean threshold.}
}
\description{
Visualization of multiple comparison DEG analysis
}
