% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA.R
\name{ORA}
\alias{ORA}
\title{ORA}
\usage{
ORA(gene_list_dir, Species, color = "qvalue")
}
\arguments{
\item{gene_list_dir}{Directory including gene list txt files}

\item{Species}{Species}

\item{color}{qvalue or p.adjust}
}
\description{
ORA
}
