% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBseq.R
\docType{data}
\name{Row_count_3conditions}
\alias{Row_count_3conditions}
\alias{ebseq}
\title{ebseq}
\usage{
data(Row_count_3conditions)
}
\arguments{
\item{Row_count_matrix}{Row Count matrix txt file (Not normalized count matrix)}
}
\description{
ebseq
}
\examples{
library(rnaseqviewer)

#pairwise DEG analysis

data(Row_count_data)
write.table(Row_count_data, file = "Row_count_data.txt", sep = "\t")
ebseq("Row_count_data.txt")

#three conditions DEG analysis

data("Row_count_3conditions")
write.table(Row_count_3conditions, file = "Row_count_3conditions.txt", sep = "\t")
ebseq("Row_count_3conditions.txt")

}
\references{
Ning Leng and Christina Kendziorski (2020). EBSeq: An R package for gene and isoform differential expression analysis of RNA-seq data.
}
