% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{Row_count_data}
\alias{Row_count_data}
\title{Row_count_data}
\format{
A data frame with 26334 rows and 4 variables
}
\usage{
data(Row_count_data)
}
\description{
example data for pairwise comparison analysis
}
