% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennd.R
\name{vennd}
\alias{vennd}
\title{Venn Diagram}
\usage{
vennd(gene_list_dir)
}
\arguments{
\item{gene_list_dir}{Directory including gene list txt files}
}
\description{
Venn Diagram
}
\examples{
library(rnaseqviewer)

data(DEG_list)
dir.create("DEG_list")
write.table(DEG_list[1], file = "DEG_list/dataset1.txt", sep = "\t", quote = FALSE)
write.table(DEG_list[2], file = "DEG_list/dataset2.txt", sep = "\t", quote = FALSE)
vennd("DEG_list")

}
\references{
Adrian Dusa (2021). venn: Draw Venn Diagrams. R package version 1.10. https://CRAN.R-project.org/package=venn
}
