% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_RNAseq_analysis.R
\name{AutoExtraction}
\alias{AutoExtraction}
\title{Automated extraction of gene set from count matrix and plot by boxplot}
\usage{
AutoExtraction(Count_matrix, Gene_set_dir)
}
\arguments{
\item{Count_matrix}{count matrix txt file}

\item{Gene_set_dir}{Directory including gene set txt files}
}
\description{
Automated extraction of gene set from count matrix and plot by boxplot
}
\examples{
library(rnaseqviewer)

data(DEG_list)
dir.create("DEG_list")
write.table(DEG_list[1], file = "DEG_list/dataset1.txt", sep = "\t", quote = F)
write.table(DEG_list[2], file = "DEG_list/dataset2.txt", sep = "\t", quote = F)
vennd("DEG_list")

data(Row_count_3conditions)
write.table(Row_count_3conditions, file = "Row_count_3conditions.txt", sep = "\t", quote = F)

AutoExtraction(Count_matrix = "Row_count_3conditions.txt", Gene_set_dir = "DEG_list/group_lists")

}
\references{
Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.7. https://CRAN.R-project.org/package=dplyr

Hadley Wickham (2021). tidyr: Tidy Messy Data. R package version 1.1.3. https://CRAN.R-project.org/package=tidyr

R. Gentleman, V. Carey, W. Huber and F. Hahne (2021). genefilter: methods for filtering genes from high-throughput experiments. R package version 1.72.1.

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Alboukadel Kassambara (2020). ggpubr: 'ggplot2' Based Publication Ready Plots. R package version 0.4.0. https://CRAN.R-project.org/package=ggpubr
}
