% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_RNAseq_analysis.R
\name{AutoExtraction}
\alias{AutoExtraction}
\title{Automated extraction of gene set from count matrix and plot by boxplot}
\usage{
AutoExtraction(Count_matrix, Gene_set)
}
\arguments{
\item{Count_matrix}{count matrix txt file}

\item{Gene_set}{Directory of gene set}
}
\description{
Automated extraction of gene set from count matrix and plot by boxplot
}
