% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_RNAseq_analysis.R
\name{GeneSetConversion}
\alias{GeneSetConversion}
\title{Gene symbol conversion from human to mouse}
\usage{
GeneSetConversion(Gene_set_dir)
}
\arguments{
\item{Gene_set_dir}{Directory including Gene_set txt files}
}
\description{
Gene symbol conversion from human to mouse
}
