% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comparison_EBseq_viewer.R
\name{multiDEG_overview}
\alias{multiDEG_overview}
\title{Visualization of multiple comparison DEG analysis}
\usage{
multiDEG_overview(
  Normalized_count_matrix,
  EBseq_result,
  EBseq_condmeans,
  Species = NULL,
  fdr = 0.05,
  fc = 2,
  basemean = 0
)
}
\arguments{
\item{Normalized_count_matrix}{Count matrix txt file (e.g. TPM count matrix.txt)}

\item{EBseq_result}{result txt file of EBseq analysis}

\item{EBseq_condmeans}{Condmeands txt file from EBseq analysis}

\item{Species}{Species}

\item{fdr}{Accepted false discovery rate for considering genes as differentially expressed}

\item{fc}{the fold change threshold. Only genes with a fold change >= fc and padj <= fdr are considered as significantly differentially expressed.}

\item{basemean}{basemean threshold.}
}
\description{
Visualization of multiple comparison DEG analysis
}
\examples{
library(rnaseqviewer)

#' #three conditions DEG analysis

data("Row_count_3conditions")
write.table(Row_count_3conditions, file = "Row_count_3conditions.txt", sep = "\t", quote = FALSE)
ebseq("Row_count_3conditions.txt")

multiDEG_overview(Normalized_count_matrix =
                  "Normalized_count_matrix_from_Cond1-vs-Cond2-vs-Cond3_EBseq.txt",
                  EBseq_result = "result_of_Cond1-vs-Cond2-vs-Cond3_EBseq.txt",
                  EBseq_condmeans ="result_of_Cond1-vs-Cond2-vs-Cond3_EBseq.condmeans",
                  Species = "human", fdr = 0.05, fc = 1.5, basemean = 0.5)

}
\references{
T Wu, E Hu, S Xu, M Chen, P Guo, Z Dai, T Feng, L Zhou, W Tang, L Zhan, X Fu, S Liu, X Bo, and G Yu. clusterProfiler 4.0: A universal enrichment tool for interpreting omics data. The Innovation. 2021, 2(3):100141

Guangchuang Yu, Li-Gen Wang, Guang-Rong Yan, Qing-Yu He. DOSE: an R/Bioconductor package for Disease Ontology Semantic and Enrichment analysis. Bioinformatics 2015 31(4):608-609

Hervé Pagès, Marc Carlson, Seth Falcon and Nianhua Li (2020). AnnotationDbi: Manipulation of SQLite-based annotations in Bioconductor. R package version 1.52.0.

Marc Carlson (2020). org.Hs.eg.db: Genome wide annotation for Human. R package version 3.12.0.

Marc Carlson (2020). org.Mm.eg.db: Genome wide annotation for Mouse. R package version 3.12.0.

R. Gentleman, V. Carey, W. Huber and F. Hahne (2021). genefilter: methods for filtering genes from high-throughput experiments. R package version 1.72.1.

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.

Alboukadel Kassambara (2020). ggpubr: 'ggplot2' Based Publication Ready Plots. R package version 0.4.0. https://CRAN.R-project.org/package=ggpubr

Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.7. https://CRAN.R-project.org/package=dplyr
}
