% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_RNAseq_analysis.R
\name{Omics_overview}
\alias{Omics_overview}
\title{Dendrogram, PCA, and mds}
\usage{
Omics_overview(Count_matrix)
}
\arguments{
\item{Count_matrix}{count matrix}
}
\description{
Dendrogram, PCA, and mds
}
