% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_RNAseq_analysis.R
\name{Omics_overview}
\alias{Omics_overview}
\title{Dendrogram, PCA, and mds}
\usage{
Omics_overview(Count_matrix)
}
\arguments{
\item{Count_matrix}{count matrix txt file}
}
\description{
Dendrogram, PCA, and mds
}
\examples{
library(rnaseqviewer)
data(Row_count_data)
write.table(Row_count_data, file = "Row_count_data.txt", sep = "\t", quote = FALSE)
deseq2("Row_count_data.txt")
Omics_overview("Normalized_count_matrix_from_Cond1-vs-Cond2_DEseq2.txt")
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Andrie de Vries and Brian D. Ripley (2020). ggdendro: Create Dendrograms and Tree Diagrams Using 'ggplot2'. R package version 0.1.22. https://CRAN.R-project.org/package=ggdendro
}
