% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA.R
\name{ORA}
\alias{ORA}
\title{Over-representation analysis (ORA)}
\usage{
ORA(gene_list_dir, Species, color = "qvalue")
}
\arguments{
\item{gene_list_dir}{Directory including gene list txt files}

\item{Species}{Species}

\item{color}{qvalue or p.adjust}
}
\description{
Over-representation analysis (ORA)
}
\examples{
library(rnaseqviewer)

data(DEG_list)
dir.create("DEG_list")
write.table(DEG_list[1], file = "DEG_list/dataset1.txt", sep = "\t", quote = FALSE)
write.table(DEG_list[2], file = "DEG_list/dataset2.txt", sep = "\t", quote = FALSE)
vennd("DEG_list")
ORA("DEG_list/group_lists", Species = "human")

}
\references{
T Wu, E Hu, S Xu, M Chen, P Guo, Z Dai, T Feng, L Zhou, W Tang, L Zhan, X Fu, S Liu, X Bo, and G Yu. clusterProfiler 4.0: A universal enrichment tool for interpreting omics data. The Innovation. 2021, 2(3):100141

Guangchuang Yu, Li-Gen Wang, Guang-Rong Yan, Qing-Yu He. DOSE: an R/Bioconductor package for Disease Ontology Semantic and Enrichment analysis. Bioinformatics 2015 31(4):608-609

Hervé Pagès, Marc Carlson, Seth Falcon and Nianhua Li (2020). AnnotationDbi: Manipulation of SQLite-based annotations in Bioconductor. R package version 1.52.0.

Marc Carlson (2020). org.Hs.eg.db: Genome wide annotation for Human. R package version 3.12.0.

Marc Carlson (2020). org.Mm.eg.db: Genome wide annotation for Mouse. R package version 3.12.0.

R. Gentleman, V. Carey, W. Huber and F. Hahne (2021). genefilter: methods for filtering genes from high-throughput experiments. R package version 1.72.1.

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.

Alboukadel Kassambara (2020). ggpubr: 'ggplot2' Based Publication Ready Plots. R package version 0.4.0. https://CRAN.R-project.org/package=ggpubr

Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.7. https://CRAN.R-project.org/package=dplyr
}
