% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example dataset.R
\docType{data}
\name{Row_count_3conditions}
\alias{Row_count_3conditions}
\title{Row_count_3conditions}
\format{
A data frame with 26334 rows and 6 variables
}
\usage{
data(Row_count_3conditions)
}
\description{
example data for 3 conditions comparison analysis
}
