% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R, R/automated_RNAseq_analysis.R
\docType{data}
\name{Row_count_data}
\alias{Row_count_data}
\alias{deseq2}
\alias{DEG_overview}
\title{deseq2}
\usage{
data(Row_count_data)

data(Row_count_data)
}
\arguments{
\item{Row_count_matrix}{Row Count matrix txt file (Not normalized count matrix)}

\item{Count_matrix}{Count matrix txt file}

\item{DEG_result}{result txt file of DEG analysis}

\item{Species}{Species}

\item{fdr}{Accepted false discovery rate for considering genes as differentially expressed}

\item{fc}{the fold change threshold. Only genes with a fold change >= fc and padj <= fdr are considered as significantly differentially expressed.}

\item{basemean}{basemean threshold}
}
\description{
deseq2

Visualization of pairwise DEG analysis
}
\examples{
library(rnaseqviewer)

data(Row_count_data)
write.table(Row_count_data, file = "Row_count_data.txt", sep = "\t")
deseq2("Row_count_data.txt")

library(rnaseqviewer)

data(Row_count_data)
write.table(Row_count_data, file = "Row_count_data.txt", sep = "\t")
deseq2("Row_count_data.txt")
DEG_overview(Count_matrix = "Normalized_count_matrix_from_Cond1-vs-Cond2_DEseq2.txt",
             DEG_result = "result_of_Cond1-vs-Cond2_DEseq2.txt",
             Species = "human")
}
\references{
Love, M.I., Huber, W., Anders, S. Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2 Genome Biology 15(12):550 (2014)

T Wu, E Hu, S Xu, M Chen, P Guo, Z Dai, T Feng, L Zhou, W Tang, L Zhan, X Fu, S Liu, X Bo, and G Yu. clusterProfiler 4.0: A universal enrichment tool for interpreting omics data. The Innovation. 2021, 2(3):100141

Guangchuang Yu, Li-Gen Wang, Guang-Rong Yan, Qing-Yu He. DOSE: an R/Bioconductor package for Disease Ontology Semantic and Enrichment analysis. Bioinformatics 2015 31(4):608-609

Hervé Pagès, Marc Carlson, Seth Falcon and Nianhua Li (2020). AnnotationDbi: Manipulation of SQLite-based annotations in Bioconductor. R package version 1.52.0.

Marc Carlson (2020). org.Hs.eg.db: Genome wide annotation for Human. R package version 3.12.0.

Marc Carlson (2020). org.Mm.eg.db: Genome wide annotation for Mouse. R package version 3.12.0.

R. Gentleman, V. Carey, W. Huber and F. Hahne (2021). genefilter: methods for filtering genes from high-throughput experiments. R package version 1.72.1.

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.

Alboukadel Kassambara (2020). ggpubr: 'ggplot2' Based Publication Ready Plots. R package version 0.4.0. https://CRAN.R-project.org/package=ggpubr

Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.7. https://CRAN.R-project.org/package=dplyr
}
